+++ src/peplum/app/peplum.py
@@ -67,6 +67,7 @@ class Peplum(EnhancedApp[None]):
                 self.theme = arguments.theme or configuration.theme
             except InvalidThemeError:
                 pass
+        self.update_keymap(configuration.bindings)
 
     def watch_theme(self) -> None:
         """Save the application's theme when it's changed."""