+++ ChangeLog.md
@@ -5,8 +5,9 @@
 **Released: WiP**
 
 - Added the ability to configure the keyboard bindings for the system
-  commands.
+  commands. ([#52](https://github.com/davep/peplum/pull/52))
 - Added `--bindings` as a command line switch.
+  ([#52](https://github.com/davep/peplum/pull/52))
 
 ## v0.5.0